extract a course function into a module
