+++ modules/gob-courses/index.js
@@ -1,0 +1,1 @@
+export {alterForEvaluation}from './alter-for-evaluation'
