+++ modules/gob-courses/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@gob/courses",
+  "version": "3.0.0-rc.2",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "lodash": "^4.17.5"
+  }
+}
